Small code changes according to conversation, namely: - renaming some methods, removing redundant/commented code, moving Sorting conversion method from NumberExt.kt to UserPreferences.kt - also fixed incorrect imports
