+++ app/src/main/java/space/taran/arknavigator/mvp/presenter/adapter/FoldersWalker.kt
@@ -1,6 +1,6 @@
 package space.taran.arknavigator.mvp.presenter.adapter
 
-import space.taran.arknavigator.mvp.model.dao.common.Preview
+import space.taran.arknavigator.ui.fragments.utils.Preview
 import space.taran.arknavigator.mvp.view.item.FileItemView
 import space.taran.arknavigator.utils.findLongestCommonPrefix
 import java.nio.file.Path
