+++ app/src/main/java/space/taran/arknavigator/mvp/view/item/FileItemViewHolder.kt
@@ -2,13 +2,11 @@ package space.taran.arknavigator.mvp.view.item
 
 import android.content.res.ColorStateList
 import android.util.Log
-import android.view.View
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.extensions.LayoutContainer
 import space.taran.arknavigator.R
 import space.taran.arknavigator.databinding.ItemFileGridBinding
-import space.taran.arknavigator.mvp.model.dao.common.Preview
+import space.taran.arknavigator.ui.fragments.utils.Preview
 import space.taran.arknavigator.utils.ITEMS_CONTAINER
 import space.taran.arknavigator.utils.imageForPredefinedIcon
 import space.taran.arknavigator.utils.loadImage
