+++ app/src/main/java/space/taran/arknavigator/utils/GlideHelper.kt
@@ -3,7 +3,7 @@ package space.taran.arknavigator.utils
 import android.widget.ImageView
 import com.bumptech.glide.Glide
 import space.taran.arknavigator.R
-import space.taran.arknavigator.mvp.model.dao.common.PredefinedIcon
+import space.taran.arknavigator.ui.fragments.utils.PredefinedIcon
 import java.nio.file.Path
 
 fun imageForPredefinedIcon(icon: PredefinedIcon): Int =
