--- app/src/main/java/space/taran/arknavigator/utils/extensions/NumberExt.kt
@@ -1,6 +1,0 @@
-package space.taran.arknavigator.utils.extensions
-
-import space.taran.arknavigator.utils.Sorting
-
-fun Int.convertToSorting(): Sorting =
-    Sorting.values()[this]
